requirements.txt
annotated-types==0.7.0
blis==1.2.0

-e file:///brainboost/packages/brainboost_desktop_package
-e file:///brainboost/packages/brainboost_configuration_package
-e file:///brainboost/packages/brainboost_data_source_logger_package
-e file:///brainboost/packages/brainboost_data_tools_pipe_pii_package
-e file:///brainboost/packages/brainboost_ocr_package

catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
easyocr==1.7.2
evdev==1.8.0
exceptiongroup==1.2.2
filelock==3.17.0
fsspec==2024.12.0
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.2
murmurhash==1.0.12
networkx==3.4.2
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==1.5.3
phonenumbers==8.13.53
pillow==11.0.0
pipdeptree==2.24.0
pluggy==1.5.0
preshed==3.0.9
presidio-anonymizer==2.0.0
presidio_analyzer==2.2.355
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
scikit-image==0.25.0
scipy==1.15.1
screeninfo==0.8.1
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
thinc==8.3.4
tifffile==2025.1.10
tldextract==5.1.3
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
